webhook/requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-datastore==1.10.0
google-cloud-ndb==0.2.1
google-cloud-profiler==1.0.9
googleapis-common-protos==1.6.0
grpcio==1.26.0
gunicorn==20.0.4
httplib2==0.15.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2019.3
redis==3.3.11
requests==2.22.0
rsa==4.0
six==1.13.0
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==0.16.0
